Restructure watch dir
