--- watch/daemon/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "daemon"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tokio-postgres = "0.7.5"
-clap = "2.33.3"
-log = "0.4.14"
-env_logger = "0.9.0"
-types = { path = "../../consensus/types" }
-eth2 = { path = "../../common/eth2" }
-tokio = { version = "1.14.0", features = ["time"] }
