--- watch/src/lib.rs
@@ -1,7 +1,0 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        // assert_eq!(2 + 2, 4);
-    }
-}
